Moved eclLogic functions and tests into IMLibrary. Added eclService to handle previous eclController functonality.
