+++ im_library/src/antlr4/ecl/ECLBuilderVisitor.js
@@ -1,7 +1,7 @@
 import ECLVisitor from "./ECLVisitor";
 import { isObjectHasKeys } from "../../helpers/DataTypeCheckers";
 import _ from "lodash";
-import { IM, SNOMED } from "@im-library/vocabulary";
+import { IM, SNOMED } from "../../vocabulary";
 
 const showLogs = false;
