+++ im_library/src/helpers/Ecl/ValidateEcl.ts
@@ -1,0 +1,10 @@
+import setupTree from "./SetupTree";
+
+export default function validateEcl(ecl: string) {
+  try {
+    const tree = setupTree(ecl);
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
