+++ im_library/src/interfaces/EclSearchRequest.ts
@@ -2,7 +2,8 @@ import { SortDirection } from "../enums/SortDirection";
 import { Query, TTIriRef } from "../interfaces/AutoGen";
 
 export interface EclSearchRequest {
-  eclQuery: Query;
+  eclQuery?: Query;
+  eclString?: string;
   includeLegacy: boolean;
   limit: number;
   statusFilter?: TTIriRef[];
